refactor: extract schema version-migration cluster into migrations.ts
Move the version-migration cluster (SCHEMA_VERSION, majorOf,
assertSchemaVersionSupported, compareVersions, needsPositionMigration,
migrateDevicePositions, clampOverRackPositions) from schemas/index.ts into
a new schemas/migrations.ts. The LayoutSchemaBase transform stays wired in
index.ts and calls the extracted helpers. index.ts re-exports SCHEMA_VERSION
and assertSchemaVersionSupported so consumers importing from $lib/schemas are
unchanged.
